src/requirements.txt
asteval==0.9.19
certifi==2020.11.8
cycler==0.10.0
Cython==0.29.21
future==0.18.2
hdbscan==0.8.26
joblib==0.17.0
kiwisolver==1.3.1
llvmlite==0.34.0
lmfit==1.0.1
matplotlib==3.3.2
numba==0.51.2
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
scikit-learn==0.23.2
scipy==1.5.4
seaborn==0.11.0
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
umap-learn==0.4.6
uncertainties==3.1.4
